feat: Create, Rename and Delete files and folders. Get file, list directory.

+++ src/FileSystemAPI.Application/Models/FileModel.cs
@@ -3,6 +3,7 @@ using FileSystemAPI.Domain.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -10,13 +11,12 @@ namespace FileSystemAPI.Application.Models
 {
     public class FileModel : AuditableEntity, IFileSystemItem
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public ItemType Type { get; set; } = ItemType.File;
         public string FileName { get; set; } = string.Empty;
         public string StoredFileName { get; set; } = string.Empty;
         public long Size { get; set; }
-        public int FolderId { get; set; }
+        public long FolderId { get; set; }
         public bool Active { get; set; }
-
     }
 }
