+++ src/FileSystemAPI.Application/Models/FolderModel.cs
@@ -1,8 +1,10 @@
 ﻿using FileSystemAPI.Application.Models.Common;
 using FileSystemAPI.Domain.Common;
+using FileSystemAPI.Domain.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -10,10 +12,10 @@ namespace FileSystemAPI.Application.Models
 {
     public class FolderModel : AuditableEntity, IFileSystemItem
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public ItemType Type { get; set; } = ItemType.Folder;
         public string FolderName { get; set; } = string.Empty;
-        public int? ParentFolderId { get; set; }
+        public long? ParentFolderId { get; set; }
         public long Size { get; set; }
         public bool Active { get; set; }
