+++ src/FileSystemAPI.Domain/Entities/File.cs
@@ -9,11 +9,11 @@ namespace FileSystemAPI.Domain.Entities
 {
     public class File : AuditableEntity
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string FileName { get; set; } = string.Empty;
         public string StoredFileName { get; set; } = string.Empty;
         public long Size { get; set; }
-        public int FolderId { get; set; }
+        public long FolderId { get; set; }
         public bool Active { get; set; }
 
         public virtual Folder Folder { get; set; } = null!;
